--- AisleNavigation/BasicMovement/README.md
@@ -1,3 +1,0 @@
-# Basic Movement
-
-Simulate back & forth movement and crude turning in Webots
