--- AisleNavigation/CarriageEndDetection/README.md
@@ -1,3 +1,0 @@
-# Carriage End Detection
-
-Using vision algorithms to detect when the robot has reached the end of a carriage.
