--- AisleNavigation/README.md
@@ -1,3 +1,0 @@
-# Aisle Navigation
-
-Milestone 1 of the Railly Clean project.
