--- CAD-Models/README.md
@@ -1,3 +1,0 @@
-# CAD Models
-
-A store of all CAD models of parts for the robot
