--- CAD-Models/Robot Requirements.md
@@ -1,9 +1,0 @@
-* Design requirements
-  * The robot must have an arm, this is to be mounted to the top of the robot
-  * The robot must have a space to hold rubbish
-  * The robot must have somewhere to store disinfectant
-  * The robot must have a connection from the disenfactant to the arm
-  * The robot must have a camera on the front and back
-  * The arm must be able to move out of the way of passengers
-  * There must be a pressure sensor on the arm
-  * There must be distance sensors on each side of the robot
