--- CAD-Models/Robot_prototype.wbo
@@ -1,194 +1,0 @@
-#VRML_OBJ R2021a utf8
-TiagoBase {
-  hidden position_0_0 260.2036853742311
-  hidden position_0_1 256.98040050070216
-  hidden position_0_2 12.789638437153423
-  hidden position_0_3 12.790008905535295
-  hidden position_0_4 31.65491424456953
-  hidden position_0_5 25.38278036122366
-  hidden position_3_0 1.1715035229883689e-06
-  hidden rotation_4 0 0.9999999999999999 0 0.12479217150351521
-  hidden position_4_0 3.498515027356914e-07
-  hidden rotation_5 0 0 1 0.5977753498514919
-  hidden position_5_0 -9.550784958041138e-07
-  hidden rotation_6 0 0 -1 2.221800955078494
-  hidden position_6_0 3.508916223785132e-07
-  hidden rotation_7 0 0 1 2.033310350891619
-  hidden position_7_0 3.7639266782052916e-07
-  hidden rotation_8 0 1 0 2.9931903760391134
-  hidden position_8_0 -7.802402592360639e-08
-  hidden position_8_1 6.562516854616344e-08
-  hidden rotation_9 -1 0 0 7.884953353001448e-08
-  hidden translation_14 1.713688855531359e-05 0.032451213642859855 -0.005702523486219135
-  hidden rotation_14 1 -8.078939978372141e-12 -3.582789557527208e-11 -0.19199999999863823
-  hidden translation_15 1.712300851730575e-05 0.08365171698253171 -2.5114860982000664e-06
-  hidden rotation_15 -0.010940060522127053 0.03870492621384009 -0.9991907944744854 -6.6090378871739126e-12
-  hidden rotation_17 1 0 0 6.664001874625056e-08
-  hidden translation_22 1.7112301710753286e-05 0.03245121531035444 0.007697483897320723
-  hidden rotation_22 1 4.7541532216822235e-12 9.094593880790001e-13 -2.9496853071815283
-  hidden translation_23 1.7136991431243587e-05 0.08365171597127308 -2.515732996144493e-06
-  hidden rotation_23 1 3.3981706374192087e-12 1.3716805483963123e-13 3.1415000000000757
-  hidden translation_33 0.00010132794445727455 -4.327232293910299e-05 0.020760374088662603
-  hidden rotation_33 0.045505925662444414 0.9976600355455502 -0.051026112970280514 9.069469916011181e-05
-  hidden translation_37 0.00011543196209151137 -0.2018319567957055 -0.0001360084779155668
-  hidden rotation_37 0.2105300651648945 0.6912151434642599 -0.691302189427567 2.726072932640712
-  hidden translation_38 -0.0013177533797993778 0.20346678258750622 8.279598694656482e-05
-  hidden rotation_38 -0.8812064940483767 0.31093561556683436 -0.3560816729614737 4.568317090922282
-  hidden translation_39 0.16920804146335883 -0.10182382943651182 -0.013395647490112916
-  hidden rotation_39 0.042641853595236485 -0.01026346594785983 0.9990377037873495 6.281534156247967
-  hidden position_39_0 857.2290944510256
-  hidden rotation_40 -0.1510689215401182 0.6989901972019442 0.6989927647410187 2.841723371910105
-  hidden translation_41 0.1692073141408067 0.10217882974322091 -0.013196039989243707
-  hidden rotation_41 0.043222604286034055 -0.01054032907643208 0.9990098637859856 6.281547497959784
-  hidden position_41_0 846.204576039103
-  hidden rotation_42 0.40383299865410494 0.6468832332899137 0.6468856094290711 3.909202728248056
-  hidden translation_43 -0.1738847688072759 -0.10157379527490872 -0.015263820113389701
-  hidden rotation_43 0.3711536312527122 -0.06334915889794948 0.9264080451263621 6.282937926859757
-  hidden position_43_0 849.7145148948127
-  hidden rotation_44 0.6105524979645148 -0.560010421019486 -0.5600124780563992 4.237874547764574
-  hidden translation_45 -0.17406047093572144 0.10249355987567017 -0.015064225780258931
-  hidden rotation_45 0.45388383886979805 0.011792556116345483 0.8909828261156654 6.282964606830066
-  hidden position_45_0 855.7356002345063
-  hidden rotation_46 -0.7103153297222576 0.49771995763896304 0.4977217858698064 1.9063651072618268
-  translation -10.077529479353085 1.5563692404840597 -6.129755683622726
-  rotation 0.5449595894237896 0.5917592230864422 0.5940034240534184 -2.1398008500599417
-  name "RaillyClean"
-  controller ""
-  bodySlot [
-    Solid {
-      translation 0 0 0.4
-      children [
-        Robot {
-          translation 0 0 0.5
-          children [
-            DEF arm IprHd6ms180 {
-              translation -0.01 0.0200031 -0.300005
-              rotation -1 0 0 -1.5708053071795867
-            }
-          ]
-        }
-        DistanceSensor {
-          translation -0.155 0 0.15
-          children [
-            DEF dSensorShape Shape {
-              appearance PBRAppearance {
-                baseColor 0 1 0
-                metalness 0
-              }
-              geometry Box {
-                size 0.01 0.05 0.05
-              }
-            }
-          ]
-          name "back distance sensor"
-          boundingObject USE dSensorShape
-          physics Physics {
-          }
-        }
-        DistanceSensor {
-          translation -0.00500018 -0.155 0.15
-          rotation 0 0 1 1.5708
-          children [
-            DEF dSensorShape Shape {
-              appearance PBRAppearance {
-                baseColor 0 1 0
-                metalness 0
-              }
-              geometry Box {
-                size 0.01 0.05 0.05
-              }
-            }
-          ]
-          name "right distance sensor"
-          boundingObject USE dSensorShape
-          physics Physics {
-          }
-        }
-        DistanceSensor {
-          translation -0.00499949 0.155 0.15
-          rotation -0.5773542691637779 -0.5773542691637779 0.5773422691581833 2.09441
-          children [
-            DEF dSensorShape Shape {
-              appearance PBRAppearance {
-                baseColor 0 1 0
-                metalness 0
-              }
-              geometry Box {
-                size 0.01 0.05 0.05
-              }
-            }
-          ]
-          name "left distance sensor"
-          boundingObject USE dSensorShape
-          physics Physics {
-          }
-        }
-        DistanceSensor {
-          translation 0.155 0 0.15
-          children [
-            DEF dSensorShape Shape {
-              appearance PBRAppearance {
-                baseColor 0 1 0
-                metalness 0
-              }
-              geometry Box {
-                size 0.01 0.05 0.05
-              }
-            }
-          ]
-          name "front distance sensor"
-          boundingObject USE dSensorShape
-          physics Physics {
-          }
-        }
-        Camera {
-          translation 0.17 0 0
-          children [
-            DEF camerashape Shape {
-              appearance PBRAppearance {
-                baseColor 1 0 0
-                metalness 0
-              }
-              geometry Box {
-                size 0.05 0.05 0.05
-              }
-            }
-          ]
-          name "front_camera"
-          boundingObject USE camerashape
-          physics Physics {
-          }
-        }
-        Camera {
-          translation -0.17 0 0
-          children [
-            DEF camerashape Shape {
-              appearance PBRAppearance {
-                baseColor 1 0 0
-                metalness 0
-              }
-              geometry Box {
-                size 0.05 0.05 0.05
-              }
-            }
-          ]
-          name "rear_camera"
-          boundingObject USE camerashape
-          physics Physics {
-          }
-        }
-        DEF Stack Shape {
-          appearance PBRAppearance {
-            metalness 0
-          }
-          geometry Box {
-            size 0.3 0.3 0.4
-          }
-        }
-      ]
-      boundingObject USE Stack
-      physics Physics {
-      }
-    }
-  ]
-}
