--- Components/README.md
@@ -1,3 +1,0 @@
-# Components
-
-Store of all python classes used by the robot.
