--- Testing/Demo 1 test cases.md
@@ -1,27 +1,0 @@
-Test cases to cover all functionality that's to be achieved by the first demo
-
-
-**Functions in project report**
-
-
-* 1 RobotStability
-  * 1.1 Robot remains upright with no arm
-  * 1.2 Robot remains upright with arm at mimimum extension
-  * 1.3 Robot remains upright with arm at maximum extension
-  * 1.4 Robot remains upright with no arm with forward simulated movement
-  * 1.5 Robot remains upright with arm at mimimum extension with forward simulated movement
-  * 1.6 Robot remains upright with arm at maximum extension with forward simulated movement
-  * 1.7 Robot remains upright with no arm with forward + sideways simulated movement
-  * 1.8 Robot remains upright with arm at mimimum extension with forward + sideways simulated movement
-  * 1.9 Robot remains upright with arm at maximum extension with forward + sideways simulated movement
-
-* 2 RobotMovement
-  * Robot moves through carriage without collisions with seats 
-  
-* 3 CarriageEndRecognition
-  * 3.1 Robot identifies carriage end from 1m
-  * 3.2 Robot identifies carriage end from 3m
-  * 3.3 Robot identified carriage end when facing 90 degrees from end
-  * 3.3 Robot identified carriage end when facing 180 degrees from end
-  * 3.4 Robot stops before collison with end
-
