+++ archive/aisle_navigation/carriage_end_detection/README.md
@@ -1,0 +1,3 @@
+# Carriage End Detection
+
+Using vision algorithms to detect when the robot has reached the end of a carriage.
