+++ archive/testing/armTest.wbt
@@ -1,0 +1,263 @@
+#VRML_SIM R2021a utf8
+WorldInfo {
+  coordinateSystem "NUE"
+}
+Viewpoint {
+  orientation 0.3402732157443384 -0.8690549007764485 -0.3591068338022047 4.621088833048081
+  position 6.167439990226379 6.758338248353503 0.5851643725230071
+}
+SolidBox {
+  translation -3.77 0.5 0
+  size 1 1 1
+}
+SolidBox {
+  translation -2.74 0.5 0.97
+  name "box(1)"
+  size 1 1 1
+}
+SolidBox {
+  translation -2.74 0.5 -1.01
+  name "box(6)"
+  size 1 1 1
+}
+SolidBox {
+  translation -1.74 0.5 -1.01
+  name "box(7)"
+  size 1 1 1
+}
+SolidBox {
+  translation -0.76 0.5 -0.94
+  name "box(8)"
+  size 1 1 1
+}
+SolidBox {
+  translation 0.22 0.5 -0.9
+  name "box(9)"
+  size 1 1 1
+}
+SolidBox {
+  translation 1.2 0.5 -0.81
+  name "box(10)"
+  size 1 1 1
+}
+SolidBox {
+  translation 2.16 0.5 -0.81
+  name "box(11)"
+  size 1 1 1
+}
+SolidBox {
+  translation -1.76 0.5 0.97
+  name "box(2)"
+  size 1 1 1
+}
+SolidBox {
+  translation -0.76 0.5 1.14
+  name "box(3)"
+  size 1 1 1
+}
+SolidBox {
+  translation 0.24 0.5 1.5
+  name "box(4)"
+  size 1 1 1
+}
+SolidBox {
+  translation 1.22 0.5 1.79
+  name "box(5)"
+  size 1 1 1
+}
+SolidBox {
+  translation 2.19 0.5 1.79
+  name "box(12)"
+  size 1 1 1
+}
+TiagoBase {
+  translation 3.40401 0.0746367 0.700824
+  rotation -0.02905469263148261 0.7051518211674953 0.7084608203283049 3.08389
+  name "RaillyClean"
+  controller "tiagoControllerCentering"
+  bodySlot [
+    Solid {
+      translation 0 0 0.4
+      children [
+        Robot {
+          translation 0 0 0.5
+        }
+        DistanceSensor {
+          translation -0.155 0 0.15
+          children [
+            DEF dSensorShape Shape {
+              appearance PBRAppearance {
+                baseColor 0 1 0
+                metalness 0
+              }
+              geometry Box {
+                size 0.01 0.05 0.05
+              }
+            }
+          ]
+          name "back distance sensor"
+          boundingObject USE dSensorShape
+          physics Physics {
+          }
+          lookupTable [
+            0 0 0
+            0.04 0 0.01
+            10 10 0.1
+          ]
+        }
+        DistanceSensor {
+          translation -0.00500018 -0.155 0.15
+          rotation 0 0 1 -1.5707953071795862
+          children [
+            DEF dSensorShape Shape {
+              appearance PBRAppearance {
+                baseColor 0 1 0
+                metalness 0
+              }
+              geometry Box {
+                size 0.01 0.05 0.05
+              }
+            }
+          ]
+          name "right distance sensor"
+          boundingObject USE dSensorShape
+          physics Physics {
+          }
+          lookupTable [
+            0 0 0
+            0.04 0 0.01
+            10 10 0.1
+          ]
+        }
+        DistanceSensor {
+          translation -0.00499949 0.155 0.15
+          rotation -0.5773542691637779 -0.5773542691637779 0.5773422691581833 2.09441
+          children [
+            DEF dSensorShape Shape {
+              appearance PBRAppearance {
+                baseColor 0 1 0
+                metalness 0
+              }
+              geometry Box {
+                size 0.01 0.05 0.05
+              }
+            }
+          ]
+          name "left distance sensor"
+          boundingObject USE dSensorShape
+          physics Physics {
+          }
+          lookupTable [
+            0 0 0
+            0.04 0 0.01
+            10 10 0.1
+          ]
+        }
+        DistanceSensor {
+          translation 0.155 0 0.15
+          rotation 0 0 1 -5.307179586466759e-06
+          children [
+            DEF dSensorShape Shape {
+              appearance PBRAppearance {
+                baseColor 0 1 0
+                metalness 0
+              }
+              geometry Box {
+                size 0.01 0.05 0.05
+              }
+            }
+          ]
+          name "front distance sensor"
+          boundingObject USE dSensorShape
+          physics Physics {
+          }
+          lookupTable [
+            0 0 0
+            0.04 0 0.01
+            10 10 0.1
+          ]
+        }
+        Camera {
+          translation 0.17 0 0
+          rotation -0.577349935856137 0.577349935856137 0.5773509358560258 -2.094395307179586
+          children [
+            DEF camerashape Shape {
+              appearance PBRAppearance {
+                baseColor 1 0 0
+                metalness 0
+              }
+              geometry Box {
+                size 0.05 0.05 0.05
+              }
+            }
+          ]
+          name "front_camera"
+          boundingObject USE camerashape
+          physics Physics {
+          }
+        }
+        Camera {
+          translation -0.17 0 0
+          children [
+            DEF camerashape Shape {
+              appearance PBRAppearance {
+                baseColor 1 0 0
+                metalness 0
+              }
+              geometry Box {
+                size 0.05 0.05 0.05
+              }
+            }
+          ]
+          name "rear_camera"
+          boundingObject USE camerashape
+          physics Physics {
+          }
+        }
+        DEF Stack Shape {
+          appearance PBRAppearance {
+            metalness 0
+          }
+          geometry Box {
+            size 0.3 0.3 0.4
+          }
+        }
+      ]
+      boundingObject USE Stack
+      physics Physics {
+      }
+    }
+  ]
+}
+Solid {
+  translation -3.27 0.471509 -0.106778
+  rotation 0.6785982482200877 0.6785982482200877 0.2810851028163115 -2.593565307179586
+  children [
+    Shape {
+      appearance PBRAppearance {
+        baseColor 0 0 1
+        metalness 0
+      }
+      geometry Box {
+        size 0.05 0.01 0.2
+      }
+    }
+    Shape {
+      appearance PBRAppearance {
+        baseColor 0 0 1
+        metalness 0
+      }
+      geometry Box {
+        size 0.2 0.01 0.05
+      }
+    }
+  ]
+}
+TexturedBackground {
+}
+TexturedBackgroundLight {
+}
+RectangleArena {
+  floorSize 10 10
+  wallHeight 0.7
+}
