+++ models/CAD Schedule.md
@@ -1,0 +1,42 @@
+* Week 3
+  * Modelling carriage (Arnav)
+    * Research specifications for size of aisle and width
+    * Create model in fusion 360
+    * Import into webots 
+  * Hardware modelling (Daniel)
+    * Implement the design that we talk about for the robot in the meeting [23/01/2021]
+    * Ideally from a set of specifications
+  * Model tables (Arnav)
+    * Research specification for their size such as height
+    *	Create models in fusion 360
+    * Add to carriage?
+    * Import into webots
+  * Robot Stability (Daniel)
+    * Research specifications for forces of trains movement
+    * Research implementing forces in webots
+  * Demo testing (both)
+    * Create test suite based off of requirements that we created
+    * Carry out testing in Webots
+    *  Record videos of testing
+* Week 4
+  * Model rubbish
+  * Model valuables
+  * Model arm
+* Week 5
+  * Model passengers
+  * Hardware modelling
+* Week 6
+  * Hardware modelling
+  * Demo testing
+* Week 7
+  * Model buttons
+  * Hardware modelling
+* Week 8
+  * Hardware modelling
+  * Demo testing
+* Week 9
+  * Model more carriages 
+  * Hardware modelling
+* Week 10
+  * Hardware modelling
+  * Demo testing
