+++ models/README.md
@@ -1,0 +1,3 @@
+# CAD Models
+
+A store of all CAD models of parts for the robot
