+++ worlds/.armTest.wbproj
@@ -1,0 +1,11 @@
+Webots Project File version R2021a
+perspectives: 000000ff00000000fd0000000300000000000001aa0000060ffc0100000002fb0000001e00480074006d006c0052006f0062006f007400570069006e0064006f00770000000000000001aa0000000000000000fb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000ffffffff0000006900ffffff00000001000004b40000045afc0200000001fb0000001400540065007800740045006400690074006f0072010000002f0000045a000000a600ffffff0000000300000c80000001b3fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c010000000000000c800000006900ffffff000007ca0000045a00000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff000000010000000200000212000006870100000002010000000100
+sceneTreePerspectives: 000000ff0000000100000002000000c00000014d0100000002010000000200
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: 2 "controllers/epuck_go_forward/epuck_go_forward.py" "controllers/carriage_end_sticker_detection/carriage_end_sticker_detection.py" "controllers/tiagoControllerCentering/tiagoControllerCentering.py"
+consoles: Console:All:All
+renderingDevicePerspectives: RaillyClean:front_camera;1;1;0;0
+renderingDevicePerspectives: RaillyClean:rear_camera;1;1;0;0
